+++ tools/azsdk-cli/Azure.Sdk.Tools.Cli/Helpers/Codeowners/ICodeownersManagementHelper.cs
@@ -5,7 +5,7 @@ using Azure.Sdk.Tools.Cli.Models;
 using Azure.Sdk.Tools.Cli.Models.AzureDevOps;
 using Azure.Sdk.Tools.Cli.Models.Responses.Codeowners;
 
-namespace Azure.Sdk.Tools.Cli.Helpers;
+namespace Azure.Sdk.Tools.Cli.Helpers.Codeowners;
 
 public interface ICodeownersManagementHelper
 {
